auth/front/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
fastapi==0.75.2
greenlet==1.1.2
Jinja2==3.1.1
paho-mqtt==1.6.1
pip==22.0.4
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
setuptools==57.0.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
wheel==0.36.2
